requirements.txt
python-dotenv
discord.py[voice]
discord
PyNaCl